requirements.txt
django
pyyaml
djangorestframework
requests
django-cors-headers
